Adding switch for firewall option and prerequisites. Closing #25

+++ helpers/tcp.psm1
@@ -20,12 +20,14 @@ function Invoke-TCP {
         Invoke-Command -ComputerName $Sender.NodeName -ScriptBlock {cmd /c "mkdir C:\Test-NetStack\tools\CTS-Traffic"} -ErrorAction SilentlyContinue
         Copy-Item C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -Destination C:\Test-NetStack\tools\CTS-Traffic -Force -ToSession $DestinationSession -ErrorAction SilentlyContinue
     }
-    
-    Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { New-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" -Direction Inbound -Program "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe" -Action Allow | Out-Null }
+
+    if ($EnableFirewallRules) {
+        Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { New-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" -Direction Inbound -Program "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe" -Action Allow | Out-Null }
+    }
 
     # CTS Traffic Rate Limit is specified in bytes/second
     $ServerLinkSpeed = $Receiver.LinkSpeed.split(" ")
-    Switch($ServerLinkSpeed[1]) {            
+    Switch($ServerLinkSpeed[1]) {
         ("Gbps") {$ServerLinkSpeedBps = [Int]::Parse($ServerLinkSpeed[0]) * [Math]::Pow(10, 9) / 8}
         ("Mbps") {$ServerLinkSpeedBps = [Int]::Parse($ServerLinkSpeed[0]) * [Math]::Pow(10, 6) / 8}
         ("Kbps") {$ServerLinkSpeedBps = [Int]::Parse($ServerLinkSpeed[0]) * [Math]::Pow(10, 3) / 8}
@@ -33,34 +35,34 @@ function Invoke-TCP {
     }
 
     $ClientLinkSpeed = $Sender.LinkSpeed.split(" ")
-    Switch($ClientLinkSpeed[1]) {              
+    Switch($ClientLinkSpeed[1]) {
         ("Gbps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) * [Math]::Pow(10, 9) / 8}
         ("Mbps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) * [Math]::Pow(10, 6) / 8}
         ("Kbps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) * [Math]::Pow(10, 3) / 8}
         ("bps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) / 8}
     }
-    
+
     $ServerRecvCounter = Start-Job `
     -ScriptBlock {
         param([string]$ServerName,[string]$ServerInterfaceDescription)
-        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
+        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
         Invoke-Command -ComputerName $ServerName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ServerInterfaceDescription)
-            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Received/sec" -MaxSamples 20 -ErrorAction Ignore 
+            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Received/sec" -MaxSamples 20 -ErrorAction Ignore
          } `
          -ArgumentList $ServerInterfaceDescription
     } `
     -ArgumentList $Receiver.NodeName,$Receiver.InterfaceDescription
-    
+
     $ServerSendCounter = Start-Job `
     -ScriptBlock {
         param([string]$ServerName,[string]$ServerInterfaceDescription)
-        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
+        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
         Invoke-Command -ComputerName $ServerName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ServerInterfaceDescription)
-            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Sent/sec" -MaxSamples 20 -ErrorAction Ignore 
+            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Sent/sec" -MaxSamples 20 -ErrorAction Ignore
          } `
          -ArgumentList $ServerInterfaceDescription
     } `
@@ -70,9 +72,9 @@ function Invoke-TCP {
     -ScriptBlock {
         param([string]$ServerName,[string]$ServerIP,[string]$ServerLinkSpeed)
         Invoke-Command -ComputerName $ServerName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ServerIP)
-            cmd /c "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -listen:$ServerIP -consoleverbosity:1 -ServerExitLimit:64 -TimeLimit:20000 -pattern:duplex" 
+            cmd /c "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -listen:$ServerIP -consoleverbosity:1 -ServerExitLimit:64 -TimeLimit:20000 -pattern:duplex"
          } `
          -ArgumentList $ServerIP
     } `
@@ -80,12 +82,12 @@ function Invoke-TCP {
 
     $ClientRecvCounter = Start-Job `
     -ScriptBlock {
-        param([string]$ClientName,[string]$ClientInterfaceDescription) 
-        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
+        param([string]$ClientName,[string]$ClientInterfaceDescription)
+        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
         Invoke-Command -ComputerName $ClientName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ClientInterfaceDescription)
-            Get-Counter -Counter "\Network Adapter($ClientInterfaceDescription)\Bytes Received/sec" -MaxSamples 20 
+            Get-Counter -Counter "\Network Adapter($ClientInterfaceDescription)\Bytes Received/sec" -MaxSamples 20
          } `
          -ArgumentList $ClientInterfaceDescription
     } `
@@ -94,31 +96,31 @@ function Invoke-TCP {
     $ClientSendCounter = Start-Job `
     -ScriptBlock {
         param([string]$ClientName,[string]$ClientInterfaceDescription)
-        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
+        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
         Invoke-Command -ComputerName $ClientName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ClientInterfaceDescription)
             Get-Counter -Counter "\Network Adapter($ClientInterfaceDescription)\Bytes Sent/sec" -MaxSamples 20
          } `
          -ArgumentList $ClientInterfaceDescription
     } `
     -ArgumentList $Sender.NodeName,$Sender.InterfaceDescription
-    
+
     $ClientOutput = Start-Job `
     -ScriptBlock {
         param([string]$ClientName,[string]$ServerIP,[string]$ClientIP,[string]$ClientLinkSpeed)
         Invoke-Command -ComputerName $ClientName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ServerIP,[string]$ClientIP,[string]$ClientLinkSpeed)
-            cmd /c "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -target:$ServerIP -bind:$ClientIP -connections:64 -consoleverbosity:1 -pattern:duplex"  
+            cmd /c "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -target:$ServerIP -bind:$ClientIP -connections:64 -consoleverbosity:1 -pattern:duplex"
          } `
          -ArgumentList $ServerIP,$ClientIP,$ClientLinkSpeed
     } `
     -ArgumentList $Sender.NodeName,$Receiver.IPAddress,$Sender.IPAddress,$ClientLinkSpeedBps
-    
+
 
     Start-Sleep 20
-                       
+
     $ServerRecv = Receive-Job $ServerRecvCounter
     $ServerSend = Receive-Job $ServerSendCounter
     $ClientRecv = Receive-Job $ClientRecvCounter
@@ -141,12 +143,12 @@ function Invoke-TCP {
     $ServerSendBitsPerSecond = [Math]::Round(($FlatServerSendOutput | Measure-Object -Maximum).Maximum, 2)
     $ClientRecvBitsPerSecond = [Math]::Round(($FlatClientRecvOutput | Measure-Object -Maximum).Maximum, 2)
     $ClientSendBitsPerSecond = [Math]::Round(($FlatClientSendOutput | Measure-Object -Maximum).Maximum, 2)
-                      
+
     Write-Verbose "Server Recv bps: $ServerRecvBitsPerSecond"
     Write-Verbose "Server Send bps: $ServerSendBitsPerSecond"
     Write-Verbose "Client Recv bps: $ClientRecvBitsPerSecond"
     Write-Verbose "Client Send bps: $ClientSendBitsPerSecond"
-                        
+
     $ServerOutput = Receive-Job $ServerOutput
     $ClientOutput = Receive-Job $ClientOutput
 
@@ -172,7 +174,9 @@ function Invoke-TCP {
     $TCPResults | Add-Member -MemberType NoteProperty -Name ReceivedPctgOfLinkSpeed -Value $ReceivedPercentageOfLinkSpeed
     $TCPResults | Add-Member -MemberType NoteProperty -Name RawData -Value $RawData
 
-    Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { Remove-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" | Out-Null }
+    if ($EnableFirewallRules) {
+        Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { Remove-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" | Out-Null }
+    }
 
     Return $TCPResults
 }
